Bugfix/modulo customers (#5)
* feat(HU-01): validaciones DTO y excepciones

* refactor(HU-01): move cliente DTO y excepciones a customers y aplicar validaciones

* refactor(HU-01): eliminar DTO redundante ClienteRequestDTO

* refactor(HU-01): adaptar excepciones para extender BusinessException y unificar manejo de errores

* refactor(architecture): reorganizar ubicación de excepciones según Clean Architecture

* refactor(architecture): reorganizar excepciones

* feat(customers): implement partial update with PATCH and validation rules

* feat(auth): implement HU11 login, logout, and security shield

* refactor(banco-digital): limpiar referencias a sprints

Limpieza general:
- Elimina referencias a "Sprint N" y "HU-N" de comentarios en código, Javadoc,
  seeds SQL, migraciones Flyway, docker-compose y .env.example; los artefactos
  quedan agnósticos al sprint para mantenerse válidos más allá de la iteración
- Elimina archivos .gitkeep en packages ya poblados (usecase, exception, model, port/in)
- Elimina console.error comentado y comentarios de código muerto en productos screen

* refactor(banco-digital): alinear modelo de dominio

Modelo de dominio — Value Objects y enums:
- Introduce TipoCuenta enum con fromNombre() case-insensitive y lanza
  TipoCuentaInvalidoException si el valor no es AHORRO/CORRIENTE
- Introduce EstadoCuenta enum; reemplaza boolean activa por EstadoCuenta estado
  en Cuenta; isActiva() delega a la comparación con el enum
- Introduce Value Objects NumeroCedula y Email en Cliente; encapsulan
  validaciones de dominio siguiendo el principio de Information Expert
- Agrega Cuenta.crearNueva() como factory method con precondiciones explícitas
  (Objects.requireNonNull, saldo inicial ZERO, fechaApertura LocalDate.now())
- CrearCuentaUseCase usa parseTipoCuenta() + Cuenta.crearNueva() en lugar de
  construir el builder manualmente con BigDecimal.ZERO inline

* refactor(banco-digital): alinear modelo de entidades.

Modelo de entidades — alineación con DDL:
- CuentaEntity: tabla cuentas → cuenta, String tipoCuenta → Short tipoCuentaId,
  boolean activa → String estado (default ACTIVA), agrega fechaApertura,
  extiende AuditableEntity, agrega @GeneratedValue(AUTO) en id
- UsuarioEntity: Long id → UUID id, Set<RolEntity> roles (ManyToMany) →
  RolEntity rol (ManyToOne), boolean bloqueado → OffsetDateTime bloqueadoHasta,
  isBloqueado() evalúa si bloqueadoHasta != null, renombra columnas (username,
  password_hash, mfa_secret, intentos_fallidos como Short)
- RolEntity: Long id con @GeneratedValue → Short id sin generación automática
- TokenRevocadoEntity: agrega Long id con @GeneratedValue, UUID usuarioId,
  OffsetDateTime revocadoAt y expiraAt; elimina Long expirationTime
- Rol y Usuario (dominio): Long id → Short id / UUID id respectivamente

Validación de entrada:
- CrearCuentaRequestDto: @NotNull en tipoCuenta → @NotBlank + @Pattern

* refactor(banco-digital): completar HUs / CustomeR
Casos de uso — implementación pendiente:
- CrearClienteUseCase: elimina throws UnsupportedOperationException (TODO Sprint 1),
  implementa flujo completo con clienteMapper.toDomain(), clienteRepository.save()
  y publicación de ClienteRegistradoEvent vía DomainEventPublisher (puerto de salida)
- ActualizarClienteUseCase: reemplaza builder manual por @With de Lombok para
  actualización inmutable campo a campo; usa ClienteNoEncontradoException en lugar
  de RuntimeException genérico; delega mapeo a ClienteMapper

* fix(auditoria): solucionar problemas de auditoría automática

- AuditableEntity: agrega @PrePersist y @PreUpdate con fallback DEFAULT_AUDITOR="SYSTEM"
  para campos createdBy/updatedBy cuando AuditorAware no está configurado

* fix(CustomerOpenApiConfig): reemplaza bean OpenAPI por OpenApiCustomizer para evitar conflicto con otras configuraciones de OpenAPI en el contexto de Spring

* test(CrearClienteUseCaseTest): reemplaza TODOs y asserts comentados por implementación real con mocks de ClienteMapper, ClienteRepositoryPort y DomainEventPublisher; verifica resultado no nulo, email y llamada a save()

---------

+++ src/main/java/com/udea/bancodigital/auth/domain/model/Rol.java
@@ -6,6 +6,6 @@ import lombok.Value;
 @Value
 @Builder
 public class Rol {
-    Long id;
+    Short id;
     String nombre;
 }
